requirements.txt
gradio>=3.50.0
openai>=0.27.0,<1.0  # Critical: Must be v0.x for API compatibility
requests>=2.31.0
pillow>=10.0.0